+++ lib/src/core/constants/constants.dart
@@ -3,6 +3,6 @@ import 'package:kurumi/src/core/themes/app_theme.dart';
 
 const kWhiteColor = Colors.white;
 const kBlackColor = Colors.black;
-const kTransparent = Colors.transparent;
+const kTransparentColor = Colors.transparent;
 final kBackgroundColor = AppTheme.background;
 final kSecondaryColor = AppTheme.secondaryColor;